Move all configs under config folder
