+++ jparest/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,1 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=uk.gov.homeoffice.digital.sas.jparest.Config
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=uk.gov.homeoffice.digital.sas.jparest.config.ApplicationConfig
