+++ jparest/src/test/java/uk/gov/homeoffice/digital/sas/jparest/config/JpaRestMvcConfigTest.java
@@ -1,4 +1,4 @@
-package uk.gov.homeoffice.digital.sas.jparest;
+package uk.gov.homeoffice.digital.sas.jparest.config;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Test;
